+++ srcs/tokenization/tokenize_utils.c
@@ -6,7 +6,7 @@
 /*   By: marsoare <marsoare@student.42porto.com>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/15 15:39:51 by marsoare          #+#    #+#             */
-/*   Updated: 2024/10/22 20:05:43 by marsoare         ###   ########.fr       */
+/*   Updated: 2024/10/22 21:12:42 by marsoare         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
